Fixed multiple dynamic relays
When trying to run a relay for multiple streams, the process would fail because of two issues:

The config "inPath" was improperly being set with the stream path. Yjr "inPath" should just be the edge server specified in the config.

Also, the onPrePlay event handler was creating a NodeRelaySession for every single task, even when the only a single onPrePlay event was fired. This would cause NMS to always return the video data for the last entry in the task list. I fixed the onPrePlay event handler to only create a NodeRelaySession for the task associated with the onPrePlayEvent.